address msvc build warnings and misspelled variable name (#635)

+++ Source/Lib/Common/Codec/EbEncDecProcess.c
@@ -257,7 +257,7 @@ static void ResetEncDec(
 #endif
     // Asuming cb and cr offset to be the same for chroma QP in both slice and pps for lambda computation
 
-    context_ptr->chroma_qp = context_ptr->qp;
+    context_ptr->chroma_qp = (uint8_t)context_ptr->qp;
 
     // Lambda Assignement
     context_ptr->qp_index = (uint8_t)picture_control_set_ptr->
@@ -296,7 +296,7 @@ static void EncDecConfigureLcu(
     context_ptr->qp = sb_qp;
 
     // Asuming cb and cr offset to be the same for chroma QP in both slice and pps for lambda computation
-    context_ptr->chroma_qp = context_ptr->qp;
+    context_ptr->chroma_qp = (uint8_t)context_ptr->qp;
     /* Note(CHKN) : when Qp modulation varies QP on a sub-LCU(CU) basis,  Lamda has to change based on Cu->QP , and then this code has to move inside the CU loop in MD */
     (void)sb_ptr;
     context_ptr->qp_index = (uint8_t)picture_control_set_ptr->
